+++ sdk/computeschedule/arm-computeschedule/src/computeScheduleClient.ts
@@ -1,27 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import {
-  createComputeSchedule,
-  ComputeScheduleContext,
-  ComputeScheduleClientOptionalParams,
-} from "./api/index.js";
-import {
-  OccurrenceExtensionOperations,
-  _getOccurrenceExtensionOperations,
-} from "./classic/occurrenceExtension/index.js";
-import { OccurrencesOperations, _getOccurrencesOperations } from "./classic/occurrences/index.js";
-import { OperationsOperations, _getOperationsOperations } from "./classic/operations/index.js";
-import {
-  ScheduledActionExtensionOperations,
-  _getScheduledActionExtensionOperations,
-} from "./classic/scheduledActionExtension/index.js";
-import {
-  ScheduledActionsOperations,
-  _getScheduledActionsOperations,
-} from "./classic/scheduledActions/index.js";
-import { TokenCredential } from "@azure/core-auth";
-import { Pipeline } from "@azure/core-rest-pipeline";
+import type { ComputeScheduleContext, ComputeScheduleClientOptionalParams } from "./api/index.js";
+import { createComputeSchedule } from "./api/index.js";
+import type { OperationsOperations } from "./classic/operations/index.js";
+import { _getOperationsOperations } from "./classic/operations/index.js";
+import type { ScheduledActionsOperations } from "./classic/scheduledActions/index.js";
+import { _getScheduledActionsOperations } from "./classic/scheduledActions/index.js";
+import type { TokenCredential } from "@azure/core-auth";
+import type { Pipeline } from "@azure/core-rest-pipeline";
 
 export { ComputeScheduleClientOptionalParams } from "./api/computeScheduleContext.js";
 
@@ -45,19 +32,10 @@ export class ComputeScheduleClient {
       userAgentOptions: { userAgentPrefix },
     });
     this.pipeline = this._client.pipeline;
-    this.occurrenceExtension = _getOccurrenceExtensionOperations(this._client);
-    this.occurrences = _getOccurrencesOperations(this._client);
-    this.scheduledActionExtension = _getScheduledActionExtensionOperations(this._client);
     this.scheduledActions = _getScheduledActionsOperations(this._client);
     this.operations = _getOperationsOperations(this._client);
   }
 
-  /** The operation groups for occurrenceExtension */
-  public readonly occurrenceExtension: OccurrenceExtensionOperations;
-  /** The operation groups for occurrences */
-  public readonly occurrences: OccurrencesOperations;
-  /** The operation groups for scheduledActionExtension */
-  public readonly scheduledActionExtension: ScheduledActionExtensionOperations;
   /** The operation groups for scheduledActions */
   public readonly scheduledActions: ScheduledActionsOperations;
   /** The operation groups for operations */